add exeption when deleting non exosting entity

+++ src/main/java/ch/kunz/joshua/flugbuchungsapi/service/AirlineService.java
@@ -45,6 +45,10 @@ public class AirlineService {
     }
 
     public void deleteById(Long id) {
-        airlineRepository.deleteById(id);
+        if (!airlineRepository.existsById(id) && id != 0) {
+            throw new IllegalArgumentException("Airline does not exist");
+        } else {
+            airlineRepository.deleteById(id);
+        }
     }
 }
